src/executables/requirements.txt
Flask==2.2.5
Pillow==9.5.0
opencv-python==4.7.0.72
numpy==1.24.3
gunicorn==20.1.0